refactor: 게시글 더보기 바텀시트 띄울 시 작성자 id도 같이 전달
