+++ src/features/community/components/PostList.tsx
@@ -6,8 +6,8 @@ import styled from 'styled-components/native';
 import { useGetPosts } from '../hooks/useGetPosts';
 import { useOpenMoreSheet } from '../hooks/useOpenMoreSheet';
 import { AllowedCategory, PostsListItem, SortParam } from '../types/postsListType';
-import MyPostModal from './MyPostModal';
-import OthersPostModal from './OthersPostModal';
+import MyPostModal from './post/MyPostModal';
+import OthersPostModal from './post/OthersPostModal';
 import PostListCard from './PostListCard';
 
 type Props = {
@@ -20,7 +20,7 @@ const PostList = ({ category, sort, scrollRef }: Props) => {
   const { data, posts, isLoading, isFetchingNextPage, isError, hasNextPage, refetch, isRefetching, fetchNextPage } =
     useGetPosts(CATEGORY_TO_BOARD_ID[category], sort);
 
-  const { openModal, closeModal, isMine, bottomSheetRef } = useOpenMoreSheet();
+  const { selectedPost, openModal, closeModal, isMine, authorId, bottomSheetRef } = useOpenMoreSheet();
 
   const renderPost: ListRenderItem<PostsListItem> = ({ item }) => (
     <PostListCard data={item} sort={sort} category={category} onOpenModal={openModal} />
@@ -53,7 +53,15 @@ const PostList = ({ category, sort, scrollRef }: Props) => {
       />
 
       <CustomBottomSheet ref={bottomSheetRef}>
-        {isMine ? <MyPostModal closeModal={closeModal} /> : <OthersPostModal closeModal={closeModal} />}
+        {selectedPost && authorId ? (
+          isMine ? (
+            <MyPostModal closeModal={closeModal} postId={selectedPost} />
+          ) : (
+            <OthersPostModal closeModal={closeModal} postId={selectedPost} authorId={authorId} />
+          )
+        ) : (
+          <></>
+        )}
       </CustomBottomSheet>
     </>
   );
