+++ src/features/community/hooks/useOpenMoreSheet.ts
@@ -5,13 +5,15 @@ import useGetVisitor from './useGetVisitor';
 // 게시글 더보기 모달 열기 훅
 export const useOpenMoreSheet = () => {
   const [selectedPost, setSelectedPost] = useState<null | number>(null);
-  const [isMine, setIsMine] = useState<boolean>(false);
+  const [authorId, setAuthorId] = useState<null | number>(null); // 게시글 작성자 ID
+  const [isMine, setIsMine] = useState<boolean>(false); // 내 게시글인지 여부
   const bottomSheetRef = useRef<BottomSheetModal>(null);
   const { data } = useGetVisitor(); // 내 id 가져오기
 
   const openModal = useCallback(
     (postId: number, authorId: number) => {
       setSelectedPost(postId);
+      setAuthorId(authorId);
 
       // 내 게시글인지 확인
       if (data) {
@@ -25,12 +27,14 @@ export const useOpenMoreSheet = () => {
 
   const closeModal = () => {
     setSelectedPost(null);
+    setAuthorId(null);
     bottomSheetRef.current?.dismiss();
   };
 
   return {
     selectedPost,
     isMine,
+    authorId,
     openModal,
     closeModal,
     bottomSheetRef,
